Add package `union`
Add all versions of the `union` (https://www.npmjs.com/package/union)
package as available under the MIT license. This is the license that is
specified in its README and LICENSE file since v0.1.0. The only thing of
note is that the licensee changed from "Nodejitsu Inc." in v0.4.3 and
earlier to "Charlie Robbins & the Contributors" in v0.4.4 and later.

Checked all versions of the package by creating an npm project with the
following dependencies and reviewing the resulting `node_modules/` dir:

    "dependencies": {
      "union-0.6.0": "npm:union@0.6.0",
      "union-0.5.0": "npm:union@0.6.0",
      "union-0.4.6": "npm:union@0.4.6",
      "union-0.4.5": "npm:union@0.4.6",
      "union-0.4.4": "npm:union@0.4.4",
      "union-0.4.3": "npm:union@0.4.3",
      "union-0.4.2": "npm:union@0.4.2",
      "union-0.4.1": "npm:union@0.4.1",
      "union-0.4.0": "npm:union@0.4.0",
      "union-0.3.8": "npm:union@0.3.8",
      "union-0.3.7": "npm:union@0.3.7",
      "union-0.3.6": "npm:union@0.3.6",
      "union-0.3.5": "npm:union@0.3.5",
      "union-0.3.4": "npm:union@0.3.4",
      "union-0.3.3": "npm:union@0.3.3",
      "union-0.3.2": "npm:union@0.3.2",
      "union-0.3.0": "npm:union@0.3.0",
      "union-0.2.1": "npm:union@0.2.1",
      "union-0.2.0": "npm:union@0.2.0",
      "union-0.1.8": "npm:union@0.1.8",
      "union-0.1.7": "npm:union@0.1.7",
      "union-0.1.6": "npm:union@0.1.6",
      "union-0.1.5": "npm:union@0.1.5",
      "union-0.1.4": "npm:union@0.1.4",
      "union-0.1.3": "npm:union@0.1.3",
      "union-0.1.2": "npm:union@0.1.2",
      "union-0.1.0": "npm:union@0.1.0"
    }

I reviewed by checking all LICENSE files have the same text and that
text matches the MIT license text (see note at the top). I also ran
`fgrep -ir license .` inside the `node_modules/` dir (after removing
the dirs of other dependencies) to look for additional license
declarations but found none.